feat: improve global toggle placement

+++ src/teams/launch-options/views/index.tsx
@@ -1,7 +1,17 @@
-import { TextField, DialogButton, SidebarNavigation, ToggleField } from '@decky/ui'
+import {
+    TextField,
+    DialogButton,
+    SidebarNavigation,
+    ToggleField,
+    showModal,
+    ConfirmModal,
+    SteamSpinner
+} from '@decky/ui'
 import { LaunchOption, useConfig } from '../../../hooks'
 import { useImmer } from 'use-immer'
 import {v4 as uuid} from 'uuid'
+import { FaPlus, FaTerminal } from "react-icons/fa"
+
 function CreateLaunchOptionForm({ configContext }: {configContext: ReturnType<typeof useConfig>}) {
     const { createLaunchOption } = configContext
     const [data, setData] = useImmer<Omit<LaunchOption, 'id'>>({
@@ -24,6 +34,11 @@ function CreateLaunchOptionForm({ configContext }: {configContext: ReturnType<ty
     }
     return (
         <div>
+            <div style={{ marginBottom: 22 }}>
+                <ToggleField label={'Enable globally'} checked={data.enableGlobally} onChange={(value) => setData((draft) => {
+                    draft.enableGlobally = value
+                })} />
+            </div>
             <div>
                 <TextField label={'Name'} style={{ width: '100%' }} value={data.name} onChange={(e) => setData((draft) => {
                     draft.name = e.target.value
@@ -39,15 +54,9 @@ function CreateLaunchOptionForm({ configContext }: {configContext: ReturnType<ty
                     draft.offCommand = e.target.value
                 })} />
             </div>
-            <div style={{ marginBottom: 22 }}>
-                <ToggleField label={'Enable globally'} bottomSeparator={'none'} checked={data.enableGlobally} onChange={(value) => setData((draft) => {
-                    draft.enableGlobally = value
-                })} />
-            </div>
             <div style={{ display: 'flex', gap: '10px' }}>
                 <DialogButton style={{ flex: 1 }}
-                              onButtonUp={(e) => { if(e.detail.button === 1) submit(); }}
-                              onMouseUp={submit}>
+                              onClick={submit}>
                     Add launch option
                 </DialogButton>
             </div>
@@ -57,11 +66,26 @@ function CreateLaunchOptionForm({ configContext }: {configContext: ReturnType<ty
 
 function UpdateLaunchOptionForm({ data, configContext }: { data: LaunchOption, configContext: ReturnType<typeof useConfig>}) {
     const {updateLaunchOption, deleteLaunchOption} = configContext
-    function deleteLO() {
-        return deleteLaunchOption(data.id)
+    function remove() {
+        return showModal(
+            <ConfirmModal
+                strTitle={`Remove launch option`}
+                strDescription={
+                    `Do you want to remove the "${data.name || 'Unnamed'}" launch option?`
+                }
+                strOKButtonText="Confirm"
+                strCancelButtonText="Cancel"
+                onOK={async () => {
+                    return deleteLaunchOption(data.id)
+                }}
+            />
+        );
     }
     return (
         <div>
+            <div style={{ marginBottom: 22 }}>
+                <ToggleField label={'Enable globally'} checked={data.enableGlobally} onChange={(value) => updateLaunchOption(data, 'enableGlobally', value)} />
+            </div>
             <div>
                 <TextField label={'Name'} style={{ width: '100%' }} value={data.name} onChange={(e) => updateLaunchOption(data, 'name', e.target.value)} />
             </div>
@@ -71,14 +95,14 @@ function UpdateLaunchOptionForm({ data, configContext }: { data: LaunchOption, c
             <div>
                 <TextField label={'Off command'} style={{ width: '100%' }} value={data.offCommand} onChange={(e) => updateLaunchOption(data, 'offCommand', e.target.value)}/>
             </div>
-            <div style={{ marginBottom: 22 }}>
-                <ToggleField label={'Enable globally'} bottomSeparator={'none'} checked={data.enableGlobally} onChange={(value) => updateLaunchOption(data, 'enableGlobally', value)} />
-            </div>
             <div style={{ display: 'flex', gap: '10px' }}>
                 <DialogButton style={{ flex: 1 }}
-                              onButtonUp={(e) => { if(e.detail.button === 1) deleteLO(); }}
-                              onMouseUp={deleteLO}>
-                    Remove launch option
+                              onClick={remove}>
+                    <div style={{
+                        color: 'oklch(63.7% 0.237 25.331)',
+                    }}>
+                        Remove launch option
+                    </div>
                 </DialogButton>
             </div>
         </div>
@@ -95,19 +119,28 @@ export function LaunchOptionsPage() {
                 background: "#0005",
             }}
         >
-            <SidebarNavigation
-                title={'Manage launch options'}
-                pages={[
-                    {
-                        title: 'Add launch option',
-                        content: <CreateLaunchOptionForm configContext={ configContext }/>,
-                    },
-                    ...configContext.config.launchOptions.map((launchOption) => ({
-                        title: launchOption.name || launchOption.id,
-                        content: <UpdateLaunchOptionForm configContext={ configContext } key={launchOption.id} data={launchOption} />
-                    }))
-                ]}
-            />
+            {
+                configContext.loading ? (
+                    <SteamSpinner width={"100%"} height={"100%"} />
+                ) : (
+                    <SidebarNavigation
+                        title={'Launch options'}
+                        showTitle={true}
+                        pages={[
+                            {
+                                icon: <FaPlus />,
+                                title: 'New launch option',
+                                content: <CreateLaunchOptionForm configContext={ configContext }/>,
+                            },
+                            ...configContext.config.launchOptions.map((launchOption) => ({
+                                icon: <FaTerminal />,
+                                title: launchOption.name || 'Unnamed',
+                                content: <UpdateLaunchOptionForm configContext={ configContext } key={launchOption.id} data={launchOption} />
+                            }))
+                        ]}
+                    />
+                )
+            }
         </div>
     )
 }
